Add functionality to button and custom size
Upon pressing button, user is prompted with an option to choose how many
squares make up the width of the sketchpad. The input is used to create
a new area of the same pixel width as before, but with internal divs
that change size based on how many are in a row.